Rename 'walker' module to 'visitor'
